Refactor: modularize timeline.tsx into smaller components with updated JSDoc
