+++ .zed/settings.json
@@ -11,6 +11,6 @@
     "node_modules",
     "*.d.ts",
     "*.config.mjs",
-    "*.lock",
-  ],
+    "*.lock"
+  ]
 }
