+++ src/components/timeline/timeline-item.tsx
@@ -1,0 +1,186 @@
+"use client";
+
+import type React from "react";
+import { useMemo } from "react";
+import { useIsMobile } from "@/hooks/use-mobile";
+import { calculateDateDiff, formatDate, parseDate } from "@/lib/date";
+import { cn } from "@/lib/utils";
+import { TimelineCard } from "./timeline-card";
+import { TimelineDot } from "./timeline-dot";
+
+/** Height for difference for each month between two timeline items in px */
+const PIXELS_PER_MONTH = 10;
+/** Minimum height for duration of a timeline item in px */
+const MIN_DURATION_HEIGHT = 150;
+/** Minimum gap height between two timeline items in px */
+const MIN_GAP_HEIGHT = 48;
+
+interface TimelineItemProps {
+	/** Starting date of the timeline item (e.g., "2023-01-01"). */
+	startDate?: string;
+	/** Ending date of the timeline item (e.g., "2023-05-01" or "Present"). */
+	endDate?: string;
+	/** Ending date of the next timeline item in the timeline sequence, used for gap spacing calculation. */
+	nextEndDate?: string;
+	/** Optional title of the timeline item. */
+	title?: string;
+	/** Optional short description of the timeline item. */
+	description?: string;
+	/** Optional detailed content or custom nodes. */
+	content?: React.ReactNode;
+	/** Optional URL path to an image associated with this timeline item. */
+	image?: string;
+	/** Optional URL slug for navigating to a detailed page. */
+	slug?: string;
+	/** Whether to render a vertical connector line extending to the next timeline item. */
+	showConnector?: boolean;
+}
+
+/**
+ * TimelineItem renders an individual event item within a timeline.
+ * It dynamically calculates its minimum height and spacing gap based on dates
+ * of the event and the subsequent event, and renders details accordingly.
+ *
+ * @param {TimelineItemProps} props - The component props.
+ * @returns {React.JSX.Element} The rendered timeline item list element.
+ */
+export function TimelineItem({
+	startDate,
+	endDate,
+	nextEndDate,
+	title,
+	description,
+	content,
+	image,
+	slug,
+	showConnector = true,
+}: TimelineItemProps): React.JSX.Element {
+	const isMobile = useIsMobile();
+
+	// Calculate heights
+	const { durationHeight, gapHeight } = useMemo(() => {
+		// Height of the job duration
+		const durationMonths = calculateDateDiff(startDate, endDate);
+		const dHeight = Math.max(
+			durationMonths * PIXELS_PER_MONTH,
+			MIN_DURATION_HEIGHT
+		);
+
+		// Height of the gap to the next job
+		let gHeight = MIN_GAP_HEIGHT;
+		if (nextEndDate && startDate) {
+			const gapMonths = calculateDateDiff(nextEndDate, startDate);
+			gHeight = Math.max(gapMonths * PIXELS_PER_MONTH, MIN_GAP_HEIGHT);
+		}
+
+		return { durationHeight: dHeight, gapHeight: gHeight };
+	}, [startDate, endDate, nextEndDate]);
+
+	let formattedStartDate = "";
+	if (startDate) {
+		formattedStartDate = formatDate(parseDate(startDate), "MMMYYYY");
+	}
+
+	let formattedEndDate = "";
+	if (endDate === "Present") {
+		formattedEndDate = "Present";
+	} else if (endDate) {
+		formattedEndDate = formatDate(parseDate(endDate), "MMMYYYY");
+	}
+
+	return (
+		<li
+			className={cn(
+				"group/item relative grid gap-6 last:pb-0",
+				isMobile ? "grid-cols-[24px_1fr]" : "grid-cols-[120px_24px_1fr]"
+			)}
+			style={{
+				minHeight: `${durationHeight}px`,
+				marginBottom: showConnector ? `${gapHeight}px` : "0px",
+			}}
+		>
+			{/* Date Column (Desktop) */}
+			{!isMobile && (
+				<div className="flex flex-col justify-between py-1 text-right">
+					<time className="font-bold text-xl leading-none">
+						{formattedEndDate}
+					</time>
+					<time className="font-bold text-xl leading-none">
+						{formattedStartDate}
+					</time>
+				</div>
+			)}
+
+			{/* Connector Column */}
+			<div className="relative flex flex-col items-center">
+				{/* Internal Line (Duration) - Changes color on hover */}
+				<span
+					className={cn(
+						"absolute top-2 bottom-2 left-1/2 w-0.5 -translate-x-1/2 bg-border transition-colors duration-300 group-hover/item:bg-muted-foreground"
+					)}
+				/>
+
+				{/* Gap Line (Connector) - Stays same color */}
+				{showConnector && (
+					<span
+						className="absolute left-1/2 w-0.5 -translate-x-1/2 bg-border"
+						style={{
+							top: "calc(100% - 8px)",
+							height: `${gapHeight + 16}px`,
+						}}
+					/>
+				)}
+
+				{/* Top Dot (End Date) */}
+				<TimelineDot />
+
+				{/* Bottom Dot (Start Date) */}
+				<div className="absolute bottom-2 left-1/2 -translate-x-1/2">
+					<TimelineDot />
+				</div>
+			</div>
+
+			{/* Content Column */}
+			<div
+				className={cn(
+					"flex flex-col",
+					isMobile ? "h-full justify-between" : "justify-center"
+				)}
+			>
+				{isMobile && (
+					<time className="mb-4 font-bold text-xl leading-none">
+						{formattedEndDate}
+					</time>
+				)}
+
+				{image ? (
+					<TimelineCard
+						content={content}
+						description={description}
+						image={image}
+						slug={slug}
+						title={title}
+					/>
+				) : (
+					<div className="flex flex-col gap-2">
+						<h3 className="font-bold text-lg leading-tight">{title}</h3>
+						{description && (
+							<p className="text-base text-muted-foreground">{description}</p>
+						)}
+						{content && (
+							<div className="whitespace-pre-line text-muted-foreground text-sm">
+								{content}
+							</div>
+						)}
+					</div>
+				)}
+
+				{isMobile && (
+					<time className="mt-4 font-bold text-xl leading-none">
+						{formattedStartDate}
+					</time>
+				)}
+			</div>
+		</li>
+	);
+}
