+++ src/components/timeline/timeline.tsx
@@ -1,58 +1,28 @@
 "use client";
 
 import type React from "react";
-import { useMemo } from "react";
-import { useIsMobile } from "@/hooks/use-mobile";
-import { formatDate } from "@/lib/date";
 import { cn } from "@/lib/utils";
-import { NavLink } from "../nav-link";
 
-/* ---------------------------------------------
- * Config
- * -------------------------------------------- */
-const PIXELS_PER_MONTH = 10;
-const MIN_DURATION_HEIGHT = 150;
-const MIN_GAP_HEIGHT = 48; // Minimum space between items
+export { TimelineItem } from "./timeline-item";
 
-/* ---------------------------------------------
- * Helpers
- * -------------------------------------------- */
-const parseDate = (dateStr?: string) => {
-	if (!dateStr || dateStr === "Present") {
-		return new Date();
-	}
-	const parts = dateStr.split(".");
-	if (parts.length === 2) {
-		return new Date(Number(parts[0]), Number(parts[1]) - 1);
-	}
-	if (parts.length === 3) {
-		return new Date(Number(parts[0]), Number(parts[1]) - 1, Number(parts[2]));
-	}
-	return new Date(dateStr);
-};
+interface TimelineProps {
+	/** Children elements (typically TimelineItem components) to render inside the timeline container. */
+	children: React.ReactNode;
+	/** Optional additional CSS classes for custom styling. */
+	className?: string;
+}
 
-const calculateMonthDiff = (start?: string, end?: string) => {
-	try {
-		const s = parseDate(start);
-		const e = parseDate(end);
-		const diff =
-			(e.getFullYear() - s.getFullYear()) * 12 + (e.getMonth() - s.getMonth());
-		return Math.max(diff, 0);
-	} catch {
-		return 0;
-	}
-};
-
-/* ---------------------------------------------
- * Timeline
- * -------------------------------------------- */
+/**
+ * Timeline component acts as a container for rendering an ordered list of
+ * timeline events or items.
+ *
+ * @param {TimelineProps} props - The component props.
+ * @returns {React.JSX.Element} The rendered timeline container.
+ */
 export function Timeline({
 	children,
 	className,
-}: {
-	children: React.ReactNode;
-	className?: string;
-}) {
+}: TimelineProps): React.JSX.Element {
 	if (!children) {
 		return (
 			<p className="py-8 text-center text-muted-foreground text-sm">
@@ -68,230 +38,5 @@ export function Timeline({
 		>
 			{children}
 		</ol>
-	);
-}
-
-/* ---------------------------------------------
- * TimelineItem
- * -------------------------------------------- */
-
-export function TimelineItem({
-	startDate,
-	endDate,
-	nextEndDate,
-	title,
-	description,
-	content,
-	image,
-	slug,
-	showConnector = true,
-}: {
-	startDate?: string;
-	endDate?: string;
-	nextEndDate?: string;
-	title?: string;
-	description?: string;
-	content?: React.ReactNode;
-	image?: string;
-	slug?: string;
-	showConnector?: boolean;
-}) {
-	const isMobile = useIsMobile();
-
-	// Calculate heights
-	const { durationHeight, gapHeight } = useMemo(() => {
-		// Height of the job duration
-		const durationMonths = calculateMonthDiff(startDate, endDate);
-		const dHeight = Math.max(
-			durationMonths * PIXELS_PER_MONTH,
-			MIN_DURATION_HEIGHT
-		);
-
-		// Height of the gap to the next job
-		let gHeight = MIN_GAP_HEIGHT;
-		if (nextEndDate && startDate) {
-			const gapMonths = calculateMonthDiff(nextEndDate, startDate);
-			gHeight = Math.max(gapMonths * PIXELS_PER_MONTH, MIN_GAP_HEIGHT);
-		}
-
-		return { durationHeight: dHeight, gapHeight: gHeight };
-	}, [startDate, endDate, nextEndDate]);
-
-	let formattedStartDate = "";
-	if (startDate) {
-		formattedStartDate = formatDate(parseDate(startDate), "MMMYYYY");
-	}
-
-	let formattedEndDate = "";
-	if (endDate === "Present") {
-		formattedEndDate = "Present";
-	} else if (endDate) {
-		formattedEndDate = formatDate(parseDate(endDate), "MMMYYYY");
-	}
-
-	return (
-		<li
-			className={cn(
-				"group/item relative grid gap-6 last:pb-0",
-				isMobile ? "grid-cols-[24px_1fr]" : "grid-cols-[120px_24px_1fr]"
-			)}
-			style={{
-				minHeight: `${durationHeight}px`,
-				marginBottom: showConnector ? `${gapHeight}px` : "0px",
-			}}
-		>
-			{/* Date Column (Desktop) */}
-			{!isMobile && (
-				<div className="flex flex-col justify-between py-1 text-right">
-					<time className="font-bold text-xl leading-none">
-						{formattedEndDate}
-					</time>
-					<time className="font-bold text-xl leading-none">
-						{formattedStartDate}
-					</time>
-				</div>
-			)}
-
-			{/* Connector Column */}
-			<div className="relative flex flex-col items-center">
-				{/* Internal Line (Duration) - Changes color on hover */}
-				<span
-					className={cn(
-						"absolute top-2 bottom-2 left-1/2 w-0.5 -translate-x-1/2 bg-border transition-colors duration-300 group-hover/item:bg-muted-foreground"
-					)}
-				/>
-
-				{/* Gap Line (Connector) - Stays same color */}
-				{showConnector && (
-					<span
-						className="absolute left-1/2 w-0.5 -translate-x-1/2 bg-border"
-						style={{
-							top: "calc(100% - 8px)",
-							height: `${gapHeight + 16}px`,
-						}}
-					/>
-				)}
-
-				{/* Top Dot (End Date) */}
-				<TimelineDot />
-
-				{/* Bottom Dot (Start Date) */}
-				<div className="absolute bottom-2 left-1/2 -translate-x-1/2">
-					<TimelineDot />
-				</div>
-			</div>
-
-			{/* Content Column */}
-			<div
-				className={cn(
-					"flex flex-col",
-					isMobile ? "h-full justify-between" : "justify-center"
-				)}
-			>
-				{isMobile && (
-					<time className="mb-4 font-bold text-xl leading-none">
-						{formattedEndDate}
-					</time>
-				)}
-
-				{image ? (
-					<TimelineImage
-						content={content}
-						description={description}
-						image={image}
-						slug={slug}
-						title={title}
-					/>
-				) : (
-					<div className="flex flex-col gap-2">
-						<h3 className="font-bold text-lg leading-tight">{title}</h3>
-						{description && (
-							<p className="text-base text-muted-foreground">{description}</p>
-						)}
-						{content && (
-							<div className="whitespace-pre-line text-muted-foreground text-sm">
-								{content}
-							</div>
-						)}
-					</div>
-				)}
-
-				{isMobile && (
-					<time className="mt-4 font-bold text-xl leading-none">
-						{formattedStartDate}
-					</time>
-				)}
-			</div>
-		</li>
-	);
-}
-
-/* ---------------------------------------------
- * TimelineDot
- * -------------------------------------------- */
-function TimelineDot() {
-	return (
-		<span
-			className={cn(
-				"relative z-10 flex h-4 w-4 items-center justify-center rounded-full border-2 border-border bg-border transition-colors duration-300 group-hover/item:bg-muted-foreground"
-			)}
-		/>
-	);
-}
-
-/* ---------------------------------------------
- * TimelineImage (Polaroid-style)
- * -------------------------------------------- */
-function TimelineImage({
-	image,
-	title,
-	description,
-	content,
-	slug,
-}: {
-	image: string;
-	title?: string;
-	description?: string;
-	content?: React.ReactNode;
-	slug?: string;
-}) {
-	const cardContent = (
-		<div className="w-full max-w-sm rotate-1 bg-card p-3 text-card-foreground shadow-lg transition-transform duration-300 group-hover/item:rotate-0">
-			<img
-				alt={title}
-				className="h-48 w-full object-cover"
-				height={"12rem"}
-				src={image}
-				width={"100%"}
-			/>
-
-			<div className="mt-3 text-center">
-				{title && (
-					<p className="font-bold font-serif text-base text-foreground">
-						{title}
-					</p>
-				)}
-
-				{description && (
-					<p className="mt-1 text-muted-foreground text-sm">{description}</p>
-				)}
-
-				{content && (
-					<div className="mt-2 whitespace-pre-line border-border border-t pt-2 text-muted-foreground text-xs">
-						{content}
-					</div>
-				)}
-			</div>
-		</div>
-	);
-
-	if (!slug) {
-		return cardContent;
-	}
-
-	return (
-		<NavLink className="block" href={`/timeline/${slug}`}>
-			{cardContent}
-		</NavLink>
 	);
 }
